requirements.txt
streamlit>=1.38.0
anthropic>=0.40.0
litellm>=1.80.0
python-dotenv>=1.1.0
pyyaml>=6.0.2
nest-asyncio>=1.6.0
